requirements.txt
chainlit
google-generativeai
requests
python-dotenv
